Update README.md
In my project, I delve into the analysis of liquor transactions in Iowa stores spanning from 2012 to 2023. 

The dataset, obtained from Google Big Query, comprises 27711028  records and encompasses 24 columns, including essential details such as date, store number, address, city, and county.

The significance of alcohol in human social lives dates back to prehistoric times. Iowa, situated in the Midwestern region of the United States, is bordered by the Mississippi River to the east and the Missouri River and Big Sioux River to the west.

For the state's liquor sales, the Iowa Department of Commerce mandates that stores selling bottled alcohol for off-the-premises consumption must possess a class "E" liquor license. This regulation, in line with most state alcohol regulatory bodies, ensures that all alcohol sales by these registered stores are meticulously logged in the Commerce department system. Notably, this data is made available as open data by the State of Iowa.

To explore the implementation of this analysis, you can directly access the code on my Google Colab.